Reworked on Code Readability, Fixed broken and buggy commands on Transfer, Deposit, and Withdraw. Removed Debt Class, refactored every single command. removed comments.
Code Readability:
After seeing how much code is unreadable because of the comments and if else statements, I decided to spend the day reworking on every single one of them ensuring this time that the code is more readable.

removed Debt to make it easier to use with Map()

Removed a lot of temporary self explanatory comments to make sure that the code isn't all full with comments that isn't necessary.

TODO:
-Add actually usable and readable comments. Using Jsdoc if possible.

-Split the App.jsx file to a few files to differentiate between command handlers etc